requirements/base.txt
boto3
Pillow
buttonshim
inky
picamera
